WIP test fixes
