+++ tests/meta/test_adaptive_random_forests.py
@@ -1,19 +1,8 @@
 from skmultiflow.data.generator.random_tree_generator import RandomTreeGenerator
 from skmultiflow.meta import AdaptiveRandomForestClassifier
+from skmultiflow.utils.utils import get_next_n_samples
 import numpy as np
 
-
-def get_next_n_samples(stream, n):
-    x_array = []
-    y_array = []
-    count = 0
-    while count < n:
-        X, y = stream.next_sample()
-        count += 1
-        x_array.append(X)
-        y_array.append(y)
-
-    return np.array(x_array).reshape((n, X.shape[1])), np.array(y_array).reshape((n))
 
 def test_adaptive_random_forests_mc():
     stream = RandomTreeGenerator(tree_random_state=112,
