+++ tests/meta/test_dynamic_weighted_majority.py
@@ -1,5 +1,5 @@
+from skmultiflow.data.generator.sea_generator import SEAGenerator
 from skmultiflow.meta import DynamicWeightedMajorityClassifier
-from skmultiflow.data import SEAGenerator
 from skmultiflow.bayes import NaiveBayes
 import numpy as np
 
@@ -24,7 +24,7 @@ def test_dynamic_weighted_majority():
             if y[0] == predictions[-1]:
                 correct_predictions += 1
         if first:
-            learner.partial_fit(X, y, classes=stream.target_values)
+            learner.partial_fit(X, y, classes=[0, 1])
             first = False
         else:
             learner.partial_fit(X, y)
